requirements.txt
certifi==2021.10.8
cfscrape==2.1.1
charset-normalizer==2.0.12
colorthief==0.2.1
file-read-backwards==2.0.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
PyAutoGUI==0.9.53
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.8
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.9
pytweening==1.0.4
pywin32==304
requests==2.27.1
six==1.16.0
urllib3==1.26.9
win10toast==0.9
